refactor : DropDownTrigger 로 리팩토링
